Disable editing of absence cells with backup care
Apparently staff in the original daycare should not mark absences during
the backup care period. Maybe supervisors could be allowed to do this,
but we don't handle this scenario at all in the UI, so it's probably
better to just disable it.

If we want to add support for supervisors later, we need to gracefully
handle the scenario where a cell has backup care *and* 0-2 absences.
Right now we don't even display absences if they exist, because the
backup care "takes over" the cell.