Execute Koski planning inside an async job
Planning can take quite a long time if there's a lot of pending Koski
work, and we don't want the calling lambda to time out just because it's
waiting for the REST response.

This will still consume 100% CPU during the planning because of
CPU-bound work and inefficient Timeline implementation, but at least it
won't blow up the lambda.