Add explicit flatpickr dependency
enduser-frontend includes transitively two different versions of
flatpickr.  We also use directly use a flatpickr CSS file, and this
worked earlier only because one of the two different versions happened
to be visible to our code. With Yarn 2 this no longer works, because we
need to have it as a dependency.