Restrict children removal from fridge family
If there is a parentship relation created, removing it should not be
done on light terms. This change will restrict unit supervisors from
removing created parentship relation. If there is a need to remove this
relation, it should be done by either `ADMIN` or `FINANCE_ADMIN` (given
they know whether the change on parentship will have an effect on the
fee decisions).