Correct daily prices for round the clock units
Alters the behavior of part months invoices for children in round the clock
units. Previously the daily prices used in part month invoices would be
calculated as a fraction of a units total operational day count, but now we use
the same daily price for all units. In the case of round the clock units, which
have more operational days that other units, it means that we only count as many
operational days in a week as a "regular" unit would have. The code assumes that
round the clock units mark absences only for the days that a child is supposed
to come to daycare. That is the way that Espoo does things currently.