Fix flakiness in SpringMvcTest
Apparently Spring MVC calls completion interceptors after the entire
response has been sent to the HTTP client. This means that it's possible
the test might be asserting state before the interceptor has closed the
database connection, resulting in a flaky test.

So, instead of using a one-shot assert, poll with a timeout.