Make sure we end up in the correct state
This is a bit of a hack, but pages with collapsibles often have
asynchronous content, and in an E2E test it's possible the click will
miss when something else on the page gets loaded and the click target
gets scrolled down. The main downside of this fix is that even though a
missed click will be retried, it's possible the first click hit
something else and changed something on the page

The real fix would be to wait until everything has been loaded so the
page is "stable", but this is much more complicated to implement in
practice.