Throw on VTJ errors
Previously Evaka would have incorrect person/guardian data because on VTJ
errors Evaka would just make a conservative guess silently. This caused for
example decisions not being sent to a child's guardians because VTJ was down
when the decisions were created and Evaka couldn't verify the guardians.
It's better to just fail in these cases because they are very rare anyway.
Evaka also assumes that whatever VTJ data we have stored in the database is
correct which obviously isn't true if the data was fetched when VTJ was down.