Replace some nulls with defaults in the raw report
The semantics were previously accidentally changed, because the old code
used manual mapping to Java primitive types (rs.getBoolean,
rs.getDouble), which return false/0.0 if the column is NULL. The new
automated mapping did not do this, because the fields were nullable in
the target data class.

So, remove nullability to clearly communicate we don't want empty cells in
the report, and explicitly coalesce to defaults in the SQL query.