Fix download attachment experience on iOS
On iOS, popups are blocked by default in Safari and confusing to use in
Chrome (suddenly a new tab and I can't go back to where I came from). So
instead of opening in a new tab (`_blank`) we now open attachments in
the same window. This way the user can now go back to the eVaka UI by
the browser's back button (or swiping).

The HTML5 download functionality has quite poor UX on iOS, as it
- on Safari it asks whether to download, but gives no indication of where the file went
- on Chrome it just looks like nothing happened
Due to this, we now open the attachments in the same window as the app
in this case as well.

EVAKA-4290