Extract translated strings for the actual vasu document
Translated strings for the actual vasu document, which can be in either
Finnish or Swedish, are now separated from the UI translations. This
enables us to translate the document separately from the UI language, so
that an employee using eVaka in Finnish can author a vasu for a Swedish
speaking citizen.