Refactor layouting of the dynamic content on vasu documents
The space between questions was added using a `<Gap>` element. This
caused issues when the last question in a section was a follow up
question, hidden in draft documents. As the last question was hidden we
saw an extra gap in that section.

Now the gaps are added using CSS (flexbox and the `gap` prop), which
solves this problem neatly.