Make sure guardians with email addresses get pedadoc notifications
The sending of pedagogical documentation notifications did not work if
one of the guardians of the child did not have an email address
configured. This makes sure that even if one guardian is missing an
email address, all the other guardians and heads of the child are sent
the notification.