Ensure session is saved before response is returned
express-session saves the session to the backing store (= Redis)
lazily and asynchronously after the HTTP response has been sent.
This introduces a possible race condition if we save something important
to the session, and assume the next request will see it immediately.
This assumption is not true, because the backing store has latency, or
might even return an error which is probably just logged and ignored
since the HTTP response has already been sent.

In practice, CSRF tokens sometimes hit this race condition...the
/auth/status endpoint saves the CSRF secret to the session, and returns
HTTP 200 response which includes a CSRF cookie. If a new request comes
in quickly enough, it will contain the CSRF cookie but the server will
not see the CSRF secret in the session, leading to a "invalid CSRF
token" error.

Fix the problem by ensuring the session is fully saved *before* we even
return a response. In logged out cases this is not needed. One slight
downside of this approach is that we always do a save, even if the
session contents have not changed. This can be improved later if
necessary, by adding some equality checks.