Make sure session is saved after req.logIn
Similar to the previously fixed race condition with CSRF secrets, the
*login sessions themselves* have a race condition: the SAML login
callback endpoint returns 200 OK before the session is saved to redis.
If the browser manages to send an /auth/status request quickly enough,
it won't see the new authenticated session, and will create *a new
unauthenticated session* which then replaces the authenticated session
id in the cookie.