Inline SOAP configuration to a single function
There's no need to have every tiny fragment of configuration as a
separate "bean" in the Spring context. This also greatly reduces
conflicts once message-service and service are merged.

SfiApiTest is also removed, since it has not been used in a long time
and is essentially dead code. The new SoapStackIntegrationTest tests
some of the same things, and doesn't require a manually set up
connection to a real environment.