EVAKA-4349 Ensures printing of vasu documents looks OK
Printing `display: flex` causes issues where text may be rendered so
that the top half of a line of text is on one page and the bottom half
on the second page. This fixes that issue by using `display: block` and
`break-inside: avoid-page` to keep relevant blocks on a single page.