Fixes flakiness in vasu e2e tests
Previously the tests would wait for the 'preview' button to be enabled
before proceeding to the preview page. This, however, caused issues if
the test was executed so quickly that the save debouncer didn't have
time to actually disable the button before the test tried to proceed.

Now we add a `data-status` attribute to the vasu container element, and
check that in our tests to determine whether saving is complete or not.