Fix flakiness
The test executor sometimes fills in a field while the previous fields
are being saved. When the PUT request returns, the status is set to
'clean' even if the document was modified while the request was
happening. Now we check whether the state has changed before setting it
to clean and if it is dirty, we initiate the debounced save again.