Ensure evaka_user row is created for paper application guardians
If a paper application is created for a guardian who has not logged in
to eVaka, they might not yet have a row in `evaka_user`. This causes
errors later when application `guardian_id` is used as income `modified_by`.

Fix future cases by adding `evaka_user` upsert, and existing cases by
upserting using a migration.