Allow deleting a person while keeping their message account
This fixes a rare case where a person has lost guardianship and would be
deleted by `InactivePeopleCleanup`, but received messages prevented the
deletion.

The message account can be kept because we add a reference to the
`evaka_user` table. This also makes it possible to remove the `active`
column; If the message account doesn't "belong to anyone", then it's
inactive.