Fix speculated occupancy preschool periods
The actual placement plan is approved in two separate steps (preschool
approval, daycare approval), so the placements are not created in one
go. However, the speculated occupancy mechanism needs to calculate
everything at once, so change the function to support this use case.

Previously the speculated occupancy mechanism called the function once
with PRESCHOOL_DAYCARE, which basically returned only the daycare period
and ignored the preschool completely.