Improve message_account table indexes
Currently we use a unique constraint per foreign key, but these indexes
are not partial indexes so they include nulls. Also, we have
separate (some_id, active) indexes, and they seem to only exist to allow
the database to do index-only scans based on (id, active) pairs without
needing to read the actual row data.

Replace both with partial indexes that include the active column, but
just include the column instead of using it as a key. Benefits:

- partial index -> no wasted space on NULLs
- active column included -> we can have the unique constraint *and*
  index-only scans based on (id, active) pairs with just one index