EVAKA-4406 Fix issue where the tooltip positioning was slightly off
The measurements of width and height of the tooltip body always used the
previous tooltip's content and applied it to the next render. So for
cases where the utilization went from e.g. 15% to 0%, causing the
tooltip body to be slightly smaller, the caret was rendered as detached
from the tooltip itself.

The above issue is fixed by specifying a fixed width/height for the
tooltip, which works when we know the contents, but is not ideal if the
content would be radically different.