EVAKA-4406 Show utilization in group selector for realtime attendances
Implements the new design for the group selector in the employee mobile.

If the realtime staff attendances feature flag is on, the current
utilization will be shown next to the name of each group in the group
selector. In other cases a `present children / total children` count
will be shown.