Use realized_placement_one in attendance reservation query
This query uses realized_placement_one to match old behaviour, which
prioritizes backup care, so a child in backup care doesn't appear in the
normal unit's results.

The old query also had a bug where a child in backup care without group
would get group info from the normal unit, and the results would contain
an id and name for a group that doesn't even exist in the backup care
unit.