Fix parameter order in DaycareFormV0 constructor
Some defaults in the constructor used this pattern:

false.takeIf { type == ApplicationType.PRESCHOOL }

The problem is that `type` refers to a field that may be set *after* it
has been referenced here, because `type` is the last parameter in the
construction. This worked in the past, but Kotlin 1.7 is stricter about
parameter order and makes this "use-before-init" order a warning.

Reordering the parameters fixes the warning.