Post SAML data in tests using urlencoded format
Previously JSON worked fine, since we had a global JSON parsing
middleware installed. Now that the global middleware no longer exists,
we *could* install the same middleware to the SAML endpoints, but in
real SAML flows the data is not posted as JSON but in urlencoded format.

So, post data in tests using that format since we don't even need to
support JSON.