Make sure reply has actually been sent
Clicking the button does not mean the reply has already been sent, and
in the tests we generally don't refresh in a loop until we see data.
This leads to race conditions where we might load the inbox on the
receiving side before the server has actually processed the click.