Fix employee login when external id changes
When employee gets a new contract or comes back to work
(or whatever) it gets a new external id in SSO for some reason but
employee number (should) stays the same.

Bad thing about this is (unline external id which seems uuid and
probably generated by system) employee number is set by human, so
there is possibility for human error to set same or wrong employee
number to employee.