Implement MVP of Action-based authorization filter
By modeling the main access control queries as QueryFragments, we can
execute them with different WHERE predicates. In this MVP we still end
up with Set<SomeId>, which is fine when the amount of ids is small.
However, this won't work universally, and eventually we want to pass
the QueryFragment all the way to the actual query that searches data.
The main challenge is that QueryFragment is a very "loose" object with no
information about the returned column names or types.