Add a button for adding a followup comment
Before this change, a new comment was created in the end automatically
when the user started to write the previous comment.

While at it, make the logic of the component logic a lot simpler. The
only downside to this is simplification that *any* change to a comment,
even if the edit is undoed, sets the editor and edit timestamp fields.