Use a cache key in access control query caching
equals/hashCode on the query itself was fine in many cases, but doesn't
work well when the query itself is parameterized. Now that queries are
defined using functions that *return QuerySql* instead of actually
executing a query, we can use the returned QuerySql object as the cache
key. This conveniently captures all information about the query,
including its parameters, so caching works even for parameterized
queries.

The only downside is slightly lower performance, because we need to
create a QuerySql twice (one for the cache check, one for the query
execution). This can be improved later by changing the Query abstraction
slightly if necessary.