Fix using the `DatePicker` calendar dialog with mouse on Safari
Remove tabindex="-1" from `<main>`. This has the effect that `<main>`
won't receive focus events, and thus fixes the premature closing of the
<DatePicker> calendar dialog on Safari.

It shouldn't break anything as tabindex="-1" means that the element can
be focused programmatically or by clicking, but not by keyboard. I don't
see why this would be required.