Cancel placements after an inserted CLUB placement
Earlier all placements were inserted such that any existing placement
would be trimmed from the ends, or split into two to make room for the
new placement. This is still the case for all other kinds of placements,
but CLUB placements only trim existing placements from the end (if they
start before the new CLUB placement) and cancel any placement (or part
of a placement) that would occur after the new CLUB placement has ended.

As an exception to this rule, PRESCHOOL, PRESCHOOL_DAYCARE, PREPARATORY
and PREPARATORY_DAYCARE placements are NOT cancelled.