Allow failing on future Flyway migrations
By default, future migrations (= in flyway_schema table but not in code)
are ignored. This makes it possible to attempt a rollback to a different
version even if the old version doesn't have include migrations.

If future migrations are not ignored, this makes it possible to detect
when data and code are out of sync, which is useful in some special
scenarios.