Don't store the placements in a vasu document before it is closed
A couple of functions used the `getVasuDocumentMaster` function to find
and update a field in the basics section of a vasu doc. Unfortunately
the `getVasuDocumentMaster` function also does a dynamic fetch of the
placements, which are _NOT_ supposed to be stored in the DB before a
document is closed. This caused the functions to prematurely store the
placement information while updating that one field in the basics
section.