Remove getPendingJobCount()
It was only used in tests, and most of the time we actually care about
the *results*, not how many planned jobs are in the database.

If the amount of jobs *is* important, it can be fetched directly using
the query, and sometimes a better metric is the amount of *executed
jobs*, which is now returned.