Use correct query to delete unclaimed async jobs
The old queries didn't use an index (because we have no index on
claimed_by = NULL) and had to search *all* async jobs. The correct way
is to use the removeUnclaimedJobs query, which also has a "completed_at
IS NULL" condition that can use a suitable index.

One of the queries also removed completed jobs, which is unnecessary
since we have a separate cleanup mechanism that handles this.