Add child+date index to attendance reservations
Currently we don't have an index like that, so in order to find
attendance reservations for one child on one date, the database needs to
search *all* reservations for the child, or *all* reservations for the
date. The exclude constraint on the table can't be used because it's
based on an expression that includes the start and end times.