Make Json binding explicit
Deriving QualifiedType from the *compile-time* type is problematic,
because if there's any inheritance hierarchy involved, the type might be
the superclass and we want the concrete class of the value.

However, this loses all type information about generics, which makes
this approach unsuitable for normal .bind usage. So, make json binding
explicit since it behaves differently.