Limit passport-saml Profile hack to one place
passport-saml typings say we can return any user type, but in reality
logout assumes it's a valid Profile. We *could* combine our fields and
the original with spread operators, but I think it's better to pick only
a subset of fields. This is also now done in only one place, instead of
having this hack copy-pasted four times