Fix loss of CSRF token
passport.js 0.6 regenerates the whole session on login and logout, which
is usually a good idea for security. However, in this case we just want
to replace the roles in an existing session and regenerating the session
leads to a new CSRF token which won't match the previously set cookie.

Since we're just updating the role information, let's just disable
session regeneration in this one location.