Fix more false corrections in the voucher value report
When the validity of a voucher value decision is shortened and it
is partially replaced with a new decision, and then the voucher
value report is frozen, if the validity of the decision is reduced
further after the freeze then the report would create corrections for
the whole time differing from the decision's original validity, instead
of just the newly reduced period. Add a left join and a where condition
which check that there is no already existing correction for some of the
months with which the validity was reduced from the original.