Attempt to improve chart scaling
react-chartjs-2 seems to be buggy, and charts can by default only shrink
and never try to expand if the window is resized. By passing width and
height explicitly (and height is *explicit* undefined!), we can activate
the responsive mode of the underlying chart.js library.

In responsive mode, the chart size is controlled by the parent, and the
chart canvas resizes automatically. Since the old behaviour is a bit
erratic, there are no obvious values for the parent size but I've used
some reasonable guesses that seem to work ok.