Remove unused settings file
Gradle automatically imports `gradle/libs.versions.toml` from the *root
project* if one exists. In the evaka core builds, the root project is
our `service` so this file is always imported automatically and the
settings file was never used.

Derived service projects still need to import the toml file in their
`settings.gradle.kts`, because they are built with the derived project
as the root project, so the automatic import doesn't find the file.