Greatly increase response read timeout in BI requests
We're streaming a very large request, and immediately after we have sent
the last byte, we start reading a response. By default the read timeout
is only 15 seconds, but if the other end is still processing the last
chunk we sent, it may take longer than that for it to respond.

There's no reason to use a small timeout here, so increase it to 5
minutes.