Send user details also in HTTP header
Send a JSON encoded AuthenticatedUser in the X-User HTTP header, in
addition to sending user data in JWT. Read the user from the header in
service, but fall back to JWT if the header does't exist.

In integration tests, send an empty JWT and use header only for the
user data.

After this has been deployed to production, reading user fields from JWT
can be removed and the empty JWT can be cached in apigw to avoid
recalculating it over and over.