Don't send feature flag to backend
Return data for all days from unit month calendar's backed, regardless
of whether the day is operation date or not. This way, the
`intermittentShiftCare` feature flag doesn't affect the response.

Also rename some fields of the response for clarity, and compute the
shift care type for each day in the backend instead of looping through
all service needs in the frontend for each day.