Use `/v1/henkilot/` for children with SSN
Prefer the `/v1/henkilot/` henkilot endpoint, which gets or creates a
henkilo, because the older `/v1/hae-henkilo/` endpoint is deprecated and
has a 500 request/day rate limit. The old endpoint is still used for
children without an SSN, because their data doesn't come from VTJ and
thus might not pass the name validation that Varda performs.