Add template_id FK index to child_document
Without this index `getTemplateSummaries` needs to do a full seq scan
for all child documents to count them. In the happy case with everything
in warm db memory this is 100+ ms which isn't terrible but is far from
optimal and will get much worse once more rows start accumulating.

With this index the counting part becomes an index only scan, and
`getTemplateSummaries` becomes a <10 ms query.