Use a lateral join to improve performance
In this query, the amount of backup cares is relatively small, so a
lateral join (= loop) works fine. The previously used normal join failed
to limit the searched set, so it basically calculated service need stats
for every backup care in the database.

This drops the query duration in various sample cases from ~300ms -> ~1ms