Combine guardian endpoints
This fixes some E2E flakiness and possible strange behaviour in real use
cases if some requests take a long time (e.g. slow network).

A better design would be to return one flat list with a boolean, because
this is what the frontend wants and we could remove some extra mapping
code. However, this would require modifying the PersonJSON type which is
used in multiple places, or duplicating a decent amount of code.