Refactor data management in invoice generation
Split the input data of invoice data to represent the four levels on
invoice generation:

- Input data for top-level invoice generation
– Input for a head of family
- Input for a child under a head of family
- Input for an invoice row under a child

These inputs are created and passed forward during invoice genration.
The classes representing the input levels are responsible for computing
derived input for their level. For example, the class for a child-level
data uses business days, operational days and absences from top-level
and head of family levels to compute child's attendance days.
Intermediate computed data is memoized for efficiency (via `by lazy`).

This makes it easier to follow invoice generation and to see which data
is specific to which level.