Handle very old Varda fee data correctly
If an entry from old system needs to be modified by eVaka, it's either
set to end the day before a new entry in eVaka data starts, or deleted
if it would be set to end before a *minimum valid end date* (which is
mandated by Varda validation rules).

This commit fixes fee data to have a separate *minimum valid end date*,
because that's how it is in Varda validation rules.