Redirect to app root if an unsupported auth endpoint is called
For some reason we sometimes receive GET requests to SAML login endpoints,
but there's no handler for the request ends up getting proxied to
evaka-service, which logs noisy errors.

It's better to just redirect to the corresponding app root in apigw,
because all valid auth routes are there so we know any other routes
don't exist and will just lead to an error later.