Fix datepicker visibility inside modals
Use `createPortal` to move date picker directly under `<body>`, and the
escape any stacking context. This makes it possible to show the picker
above all other content in all cases.

In addition to window resize, also track the scroll events of all parent
elements to correctly position the picker when the date input moves.

The picker now also automatically shows above the date input if there's
not enough room under it.

react-focus-lock is used to restore focus when the picker is closed.